+++ Tests/test_glyph_render.py
@@ -176,7 +176,7 @@ def test_vectorBackends(backendName, imageSuffix):
     # - CoreGraphics PDFs are weirdly different while looking the same
     # assert expectedPath.read_bytes() == outputPath.read_bytes()
     diff = compareImages(expectedPath, outputPath)
-    assert diff < 0.00012, diff
+    assert diff < 0.00012, f"diff = {diff} ({expectedPath}, {outputPath})"
 
 
 def test_recursive():
